[modify] 機能を分解
